requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
requests
python-dotenv
google-cloud-storage
apache-airflow-providers-google
pymongo[srv]
certifi